Improve the basic usage example of logging (slightly)
By removing the responsibility of instantiating the log channel from the ping handler.

It would be cumbersome and inefficient if we instantiated this logging service in multiple places within the application.

To begin building a better solution, we can use a factory to build the logging service and declare it as the default implementation of the PSR-3 logger interface in the config provider.

The container now knows how to build this logging service, so we can inject it as a dependency into the ping handler (and any other services that require it).