Improve logging basic usage example by using basic implementation of [PSR-3](https://www.php-fig.org/psr/psr-3/) LoggerAwareInterface
This basic implementation is already provided by the official [PHP-FIG log repository](https://github.com/php-fig/log)

Using an initializer we can automatically inject the logger as a dependency (using setter method) into classes implementing the LoggerAwareInterface.

It's unwise to overuse initializers because they introduce a slight performance hit (initializers run everytime a service is loaded from the container). However, the chances are we will need to widely use logging within the application so in this instance using an initializer feels appropriate.

Our ping handler is now much tidier, as now we simply need to implement the LoggerAwareInterface in order to have access to logging capabilities.