Install PSR-3 compliant test logger as dev dependency
The [psr/log package](https://github.com/php-fig/log) used to have not only the Interface for PSR-3 Logger, but also actual implementations of the interface like the TestLogger.

The TestLogger could be used as mock for any PSR-3 Logger in your test cases.

However, from v3 the TestLogger was removed, so that the psr/log package would focus solely on the Interface.

Thanks to the community there is already a replacement for the TestLogger:

So just install the https://github.com/colinodell/psr-testlogger and replace the definition.

Reference: https://nerdpress.org/2022/03/04/class-psrlogtesttestlogger-not-found/