Add a domain request to create a user
The [API documentation](docs/openapi.yaml) specifies that API clients will provide an email and password in the API request body to create a user.

Therefore, a good design for the (create user) domain request is to use an email/password to instantiate the request and validate these fields are valid.

I installed a validation library [laminas-validator](https://docs.laminas.dev/laminas-validator/), which abstracts validation logic, to make validation quick and easy.