Add starting point for the domain to create a user
This is a minimum viable product (MVP). I have added todo notes to revisit this domain in follow-up tasks to (a) introduce a rich user model (b) decouple the domain from Doctrine (c) dispatch a user created event.

For now, it's a starting point that should work, according to the unit tests.