By default, use ReflectionBasedAbstractFactory as an abstract factory
When the service manager has not been instructed on how to build a service, by default it will now use the ReflectionBasedAbstractFactory to build the service.

Given that the class uses reflection, there will be a performance hit, but for now the convenience of not having to explicitly register a factory for each application service outweighs this con.

More info - https://matthewsetter.com/easy-instantiation-with-reflection-factory/