Add starting point for a user service
This service composes using the Doctrine entity manager as a dependency, so that actions can be performed against the user repository.

It allows the Doctrine entity manager to be coupled to this service layer, rather than it having to be coupled to every application domain where it may be useful.

This would make it easier to switch out Doctrine ORM for alternative ORMs in the future, should it ever be necessary.